requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
psycopg2-binary==2.9.9